Move focusRing back to includes
