+++ packages/Stylers/src/includes.js
@@ -51,6 +51,45 @@ export const visuallyHidden = css`
   width: 1px !important;
 `;
 
+// Focus Indicators
+
+export const focusRing = (inset = false) => {
+  return inset === true
+    ? css`
+        box-shadow: ${tokens.highlight.active.withBorder.insetBoxShadow};
+        outline: none;
+      `
+    : css`
+        box-shadow: ${tokens.highlight.active.withBorder.boxShadow};
+        outline: none;
+      `;
+};
+
+focusRing.bordered = (inset = false) => {
+  return inset === true
+    ? css`
+        border-color: ${tokens.highlight.active.noBorder.borderColor};
+        box-shadow: ${tokens.highlight.active.noBorder.insetBoxShadow};
+        outline: none;
+      `
+    : css`
+        border-color: ${tokens.highlight.active.noBorder.borderColor};
+        box-shadow: ${tokens.highlight.active.noBorder.boxShadow};
+        outline: none;
+      `;
+};
+
+focusRing.subtle = (inset = false) => {
+  return inset === true
+    ? css`
+        outline: ${tokens.highlight.active.nonInteractive.outline};
+        outline-offset: -2px;
+      `
+    : css`
+        outline: ${tokens.highlight.active.nonInteractive.outline};
+      `;
+};
+
 //
 // Forms
 //
